+++ backend/api/pyproject.toml
@@ -1,0 +1,25 @@
+[project]
+name = "backend"
+version = "0.1.0"
+description = ""
+authors = [
+    {name = "jonanfu",email = "jonanfu@gmail.com"}
+]
+readme = "README.md"
+requires-python = ">=3.9,<3.13"
+dependencies = [
+    "supabase (>=2.22.0,<3.0.0)",
+    "uvicorn[standard] (>=0.37.0,<0.38.0)",
+    "fastapi (>=0.119.0,<0.120.0)",
+    "python-dotenv (>=1.1.1,<2.0.0)"
+
+]
+
+[tool.poetry]
+packages = [
+    { include = "api", from = "src" },
+]
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
