--- backend/src/rasa/data/nlu.yml
@@ -1,16 +1,0 @@
-version: "3.1"
-
-nlu:
-- intent: consultar_lista_pokemon
-  examples: |
-    - muéstrame todos los pokémon
-    - lista de pokémon
-    - quiénes son todos los pokémon
-    - quiero ver los pokémon disponibles
-
-- intent: consultar_pokemon_especifico
-  examples: |
-    - información de [pikachu](pokemon_name)
-    - detalles de [charizard](pokemon_name)
-    - dame datos de [bulbasaur](pokemon_name)
-    - cómo es [squirtle](pokemon_name)
