--- backend/src/rasa/data/stories.yml
@@ -1,12 +1,0 @@
-version: "3.1"
-
-stories:
-- story: obtener lista de pokémon
-  steps:
-  - intent: consultar_lista_pokemon
-  - action: action_get_all_pokemon
-
-- story: obtener detalles de un pokémon específico
-  steps:
-  - intent: consultar_pokemon_especifico
-  - action: action_get_pokemon_detail
