--- backend/src/rasa/endpoints.yml
@@ -1,2 +1,0 @@
-action_endpoint:
-  url: "http://localhost:5055/webhook"
