+++ front/app/api/openapi/route.ts
@@ -1,0 +1,7 @@
+import { NextResponse } from 'next/server'
+import { openApiSpec } from '@/lib/openapi' 
+
+export async function GET() {
+  return NextResponse.json(openApiSpec)
+}
+
