+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "@scalar/nextjs-api-reference": "^0.8.22"
+  }
+}
